Create README.md
A Blazor project is developed for user creation, login, and authentication using .NET Core 8. The project consists of three layers and is WebAssembly (WASM) enabled. It's important to note that WASM in .NET Core 8 has undergone significant changes. To ensure compatibility with the previous template and to simplify the project setup, it needs to be emulated. I have already done this emulation. Additionally, JWT is used for the TOKEN section.